backend/catalog/serializers.py
from rest_framework import serializers

from .models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = [
            "id",
            "name",
            "kind",
            "is_active",
            "created_at",
            "updated_at",
        ]
